Move Traverse in its own module
